fix(wm): handle moving windows to/from floating workspaces
This commit fixes the issue related to moving windows to/from a floating
workspace to a tiled workspace. Previously the start of the move would
be ignored however when moving back from a tiled workspace since it
didn't know about the existance of that window it would also "move" that
workspace focused tiled window without physically moving it, leaving it
in a weird state that seemed like it was unmanaged.
This commit changes the way this mouse moves are handled and now also
handles moving `floating_windows` and even monocle or maximized windows.