feat(bar): optional workspaces on Komorebi widget
This commit makes the `workspaces` on `Komorebi` widget optional. This
way it allows adding the `workspaces` on one Alignment and the
`focused_window` on another one, for example.

For some reason if you had some widget on the `center_widgets` with only
icons, like the `focused_window` with just icons or the `workspaces`
with just icons and nothing else, EGUI wouldn't show the icon properly.
I believe that on the first frame the icons might not have been loaded
yet, so it set the height of the `center_widgets` Area to 0 and
afterwards it wouldn't grow to fit the icons unless we've made some
config change for example.

To fix this we've made the icons have fixed size which is equal to the
font size. For this it was required that the font size was added to the
`RenderConfig` so that the widgets would have access to it when
rendering.