fix(client): expose `new` function from `Monitor`
Currently if one wants to create a new `Monitor` using the data from
`win32_display_data` it needs to be done using the `new` function only
accessible through the `komorebi::monitor` module. This commit adds a
`new` function to the `Monitor` itself which calls the previously
mentioned one, meaning that anyone with access to the `komorebi-client`
will have access to this function as well.