fix(monitor_reconciliator): register more events
This commits adds aa few more events that can trigger a
`DisplayConnectionChange` event. Some of these events are redundant and
after a display is disconnected/reconnected it emits multiple
`DisplayConnectionChange` events, however when trying to remove them to
have just one it stopped behaving as it should as if it was missing the
update while having them duplicated it works properly, so it appears to
be better to keep them for now since the duplicated events will exit
early as soon as they see that the monitors count match (on the first
event the counts don't match so it adds/removes the monitor and the
following events see that the counts match) so it shouldn't be an issue.