refactor(wm): refactor alt-tab reconciliation
This commit changes the way the alt-tab reconciliation is done. It no
longer uses the `workspace_reconciliator` with the sending of
notifications. Instead there is a simple function that checks if the
shown/uncloaked window is already handled by komorebi and if it is it
will check if it is on an unfocused workspace and/or if it is an
unfocused window of a stacked container, returning the
monitor/workspace index pair of said window. If we get this pair then we
perform the reconciliation immediately by focusing the monitor,
workspace, container and window indices corresponding to that window.