fix(wm): add write-timeout for command reply
This commit adds a write timeout to the `process_command` unix stream.
Previously it had just the read timeout. A write timeout was added to
the `komorebi-client` functions that send the messages or queries.
Although those seem to act only on the sending side of the connection,
so this commit adds a write timeout on the receiving side of komorebi so
that when it writes back to the socket to reply to the query it doesn't
get stuck writing. This is aimed to potentially fix an issue where
sometimes the `process_command` thread would get stuck when komorebi
restarted while other 3rd party programs like YASB were already running.
The `state` query emitted by YASB right after a connection is accepted
was sometimes making the `process_command` get stuck reading or writing
to the socket.